Update CSS style sheet to add more functionality and better responsiveness
Added comments to a few features to explain more to come.

added more functionality to hero image so it scales better and faces stay visible.

Added margin auto and max-width of 1200px to all about sections on index.html and offerings.html

fixed a hero-img-out to hero-img as I put the wrong one in there around lines 457-469